+++ src/front/filters.js
@@ -1,0 +1,3 @@
+export function blankLine(val) {
+  return val.trim() === '' ? '&nbsp;' : val
+}